Moved /addresses/active to separate file
